pyproject.toml
[project]
name = "usycal"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright (>=1.56.0,<2.0.0)",
    "icalendar (>=6.3.2,<7.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pytest (>=9.0.1,<10.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
